refactor: api 함수 이름 cart를 넣어서 장바구니 요청이라는 의미 강조
