+++ src/api/delete/deleteCartItem.ts
@@ -1,8 +1,8 @@
 import axiosInstance from '../../utils/axios';
 import { URLS } from '../../constants/constants';
 
-const deleteItem = async (cartId: number) => {
+const deleteCartItem = async (cartId: number) => {
   await axiosInstance.delete(URLS.DELETE_ITEMS(cartId));
 };
 
-export default deleteItem;
+export default deleteCartItem;
